core: RK operand format — drop ADDK/MULK, 9-bit const/reg C field
Lua-style RK encoding: C < 256 selects a register, C >= 256 a constant
pool index (0..255). All arithmetic, bitwise, shift and comparison ops
read C through this scheme, so literals fold into a single instruction
with no LOADK. Removes OP_ADDK/OP_MULK and the asymmetric K duplication.
ABC layout becomes opcode 7 + A 8 + B 8 + C 9; AsBx/ABx gain a 17-bit
field. Fixes jump-patch bit positions, RK operand truncation and
shift/bitwise operand reads.

+++ src/chunk.h
@@ -51,7 +51,7 @@ void chunk_free(Chunk *chunk);
 int  chunk_write(Chunk *chunk, uint32_t inst, int line);
 
 /* Convenience emitters using encoding macros from opcode.h */
-int  chunk_emit_ABC (Chunk *chunk, int line, OpCode op, uint8_t a, uint8_t b, uint8_t c);
+int  chunk_emit_ABC (Chunk *chunk, int line, OpCode op, uint8_t a, uint8_t b, uint32_t c);
 int  chunk_emit_ABx (Chunk *chunk, int line, OpCode op, uint8_t a, uint16_t bx);
 int  chunk_emit_AsBx(Chunk *chunk, int line, OpCode op, uint8_t a, int32_t sbx);
